+++ addons/yat/src/commands/Extensible.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Text;
 using YAT.Attributes;
 using YAT.Helpers;
+using YAT.Interfaces;
 
 namespace YAT.Commands
 {