+++ addons/yat/src/commands/builtin/History.cs
@@ -1,7 +1,9 @@
 using System.Linq;
 using System.Text;
 using YAT.Attributes;
+using YAT.Enums;
 using YAT.Helpers;
+using YAT.Interfaces;
 
 namespace YAT.Commands
 {