+++ addons/yat/src/commands/builtin/List.cs
@@ -2,6 +2,8 @@ using System;
 using System.Linq;
 using System.Text;
 using YAT.Attributes;
+using YAT.Enums;
+using YAT.Interfaces;
 
 namespace YAT.Commands
 {