+++ addons/yat/src/commands/builtin/Man.cs
@@ -1,5 +1,7 @@
 using YAT.Attributes;
+using YAT.Enums;
 using YAT.Helpers;
+using YAT.Interfaces;
 
 namespace YAT.Commands
 {