+++ addons/yat/src/commands/builtin/Restart.cs
@@ -1,4 +1,6 @@
 using YAT.Attributes;
+using YAT.Enums;
+using YAT.Interfaces;
 
 namespace YAT.Commands
 {