+++ addons/yat/src/commands/builtin/Set.cs
@@ -1,5 +1,7 @@
 using YAT.Attributes;
+using YAT.Enums;
 using YAT.Helpers;
+using YAT.Interfaces;
 using YAT.Overlay.Components.Terminal;
 
 namespace YAT.Commands