+++ addons/yat/src/commands/builtin/View.cs
@@ -1,6 +1,8 @@
 using System;
 using Godot;
 using YAT.Attributes;
+using YAT.Enums;
+using YAT.Interfaces;
 using YAT.Overlay.Components.Terminal;
 
 namespace YAT.Commands