+++ addons/yat/src/commands/builtin/Watch.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Threading;
 using YAT.Attributes;
+using YAT.Enums;
 using YAT.Helpers;
+using YAT.Interfaces;
 using YAT.Overlay.Components.Terminal;
 
 namespace YAT.Commands