+++ addons/yat/src/enums/CommandResult.cs
@@ -1,0 +1,13 @@
+namespace YAT.Enums
+{
+	/// <summary>
+	/// Represents the result of executing a command.
+	/// </summary>
+	public enum CommandResult
+	{
+		Success,
+		Failure,
+		InvalidArguments,
+		InvalidCommand,
+	}
+}