+++ script_templates/Node/ExtensibleCommand.cs
@@ -1,13 +1,19 @@
 // meta-description: Base template for extensible YAT commands.
 
 using _BINDINGS_NAMESPACE_;
-using YAT;
+using YAT.Attributes;
+using YAT.Enums;
+using YAT.Interfaces;
 
 namespace YAT.Commands
 {
 	[Command("_CLASS_", "Lorem ipsum dolor sit amet.", "[b]Usage[/b]: _CLASS_")]
 	public partial class _CLASS_ : Extensible, ICommand
 	{
+		public YAT Yat { get; set; }
+
+		public _CLASS_(YAT Yat) => this.Yat = Yat;
+
 		public CommandResult Execute(YAT yat, params string[] args)
 		{
 			var option = args[1];